+++ _inc/client/components/jetpack-termination-dialog/README.md
@@ -11,7 +11,7 @@ This dialog is used to confirm with the user if they want to disconnect from Jet
 Roughly adapated form `components/connect-button/index.jsx`
 
 ```jsx
-import JetpackDisconnectDialogModal from 'components/jetpack-termination-dialog/modal';
+import JetpackDisconnectModal from 'components/jetpack-termination-dialog/disconnect-modal';
 
 <Button onClick={ this.setState( { showModal: ! this.state.showModal } ) } />
